Add new feature to main.cpp

+++ Main/funkcje.h
@@ -10,16 +10,16 @@ void dodaj(std::vector<Czytelnik>& czytelnicy, size_t& last_id);
 
 void wyswietl(std::vector<Czytelnik>& czytelnicy);
 
-void edytuj(std::vector<Czytelnik>& czytelnicy, int index);
+void edytuj(std::vector<Czytelnik>& czytelnicy, size_t index);
 
-void usun(std::vector<Czytelnik>& czytelnicy, int index);
+void usun(std::vector<Czytelnik>& czytelnicy, size_t index);
 
 void dodaj(std::vector<Ksiazka>& ksiazki);
 
 void wyswietl(std::vector<Ksiazka>& ksiazki);
 
-void edytuj(std::vector<Ksiazka>& ksiazki, int index);
+void edytuj(std::vector<Ksiazka>& ksiazki, size_t index);
 
-void usun(std::vector<Ksiazka>& ksiazki, int index);
+void usun(std::vector<Ksiazka>& ksiazki, size_t index);
 
 #endif
